data_analysis/requirements.txt
pandas
psycopg2-binary
vaderSentiment
scikit-learn
torch
joblib
numpy